bugprone-implicit-widening-of-multiplication-result fixes for CartGrid.C, MeshBlock.h and MeshBlock.C

+++ src/CartGrid.C
@@ -216,15 +216,17 @@ void CartGrid::search(double* x, int* donorid, int npts)
                 }
             }
         }
-        if (myid == 2 && abs(x[3 * i] - 0.739573) < 1e-5 &&
+        if (myid == 2 && abs(x[static_cast<int>(3 * i)] - 0.739573) < 1e-5 &&
             abs(x[3 * i + 1] + 0.259310) < 1e-5 &&
             abs(x[3 * i + 2] + 0.639614) < 1e-5) {
             printf(
-                "%d %d %f %f %f %d\n", myid, i, x[3 * i], x[3 * i + 1],
-                x[3 * i + 2], donorid[i]);
+                "%d %d %f %f %f %d\n", myid, i, x[static_cast<int>(3 * i)],
+                x[3 * i + 1], x[3 * i + 2], donorid[i]);
         }
         if (donorid[i] == -1) {
-            printf("%d %f %f %f\n", myid, x[3 * i], x[3 * i + 1], x[3 * i + 2]);
+            printf(
+                "%d %f %f %f\n", myid, x[static_cast<int>(3 * i)], x[3 * i + 1],
+                x[3 * i + 2]);
         }
     }
     // printf("CartGrid::search Processor %d located %d of %d
